requirements.txt
# UI
streamlit>=1.39.0

# Parsing
pdfplumber>=0.11.4
pillow>=10.4.0

# OCR (requires OS tesseract)
pytesseract>=0.3.10

# NLP / Entities
spacy>=3.7.5

# Utilities
tqdm>=4.66.4
rapidfuzz>=3.9.6
